% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsa_fit.R
\name{dlsa.fit}
\alias{dlsa.fit}
\title{Distributed Least square approximation}
\usage{
dlsa.fit(
  formula,
  data,
  K,
  fit_fun,
  lasso_fun = 1,
  ind = NULL,
  intercept = 1,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{K}{number of workers}

\item{fit_fun}{user-submitted fitting function}

\item{lasso_fun}{logical. Should the lars result be included in the final result?}

\item{ind}{the rule to split the data to workers}

\item{intercept}{logical. Should an intercept be included in the null model?}

\item{...}{optional arguments to fit_fun}
}
\value{
A list of the distributed least squares estimator and covariance matrix
}
\description{
Distributed Least square approximation
}
\examples{
# Not Run
}
