% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsa_logistic.R
\name{dlsa.logistic}
\alias{dlsa.logistic}
\title{Distributed Least square approximation of logistic regression}
\usage{
dlsa.logistic(
  formula,
  data,
  K,
  lasso_func = 1,
  alpha = 0.5,
  init.beta = NULL,
  iter.max = NULL,
  ind = NULL,
  intercept = 1
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{K}{number of workers}

\item{lasso_func}{logical. Should the lars result be included in the final result?}

\item{alpha}{the critical point to obtain Yhat from Fitted result}

\item{init.beta}{initial beta values for the logistic regression}

\item{iter.max}{max iterate times for the logistic regression}

\item{ind}{the rule to split the data to workers}

\item{intercept}{logical. Should an intercept be included in the null model?}
}
\value{
A list of the distributed least squares estimator and covariance matrix
}
\description{
Distributed Least square approximation of logistic regression
}
\examples{
# Not Run
}
