% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsa_plot.R
\name{dlsa.plot}
\alias{dlsa.plot}
\title{Plot for the results}
\usage{
dlsa.plot(plot_fun, roc = 0, Y = NULL, p = NULL, ...)
}
\arguments{
\item{plot_fun}{user-submitted plotting function}

\item{roc}{logical. Should the roc curve be included in the final result?}

\item{Y}{a factor, numeric or character vector of responses, typically encoded with 0 (controls) and 1 (cases)}

\item{p}{a numeric or ordered vector of the same length than response, containing the predicted probability of each observation.}

\item{...}{optional arguments to plot_fun}
}
\value{
A list of the plots
}
\description{
Plot for the results
}
\examples{
# Not Run
}
