% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsa_pred.R
\name{dlsa.pred}
\alias{dlsa.pred}
\title{Use distributed methods to predict dependent variables based on estimated parameters}
\usage{
dlsa.pred(X, theta, K, pred_fun, ind = NULL, ...)
}
\arguments{
\item{X}{a symbolic description of the model to be fitted.}

\item{theta}{a vector containing the estimated parameters of the model.}

\item{K}{number of workers}

\item{pred_fun}{user-submitted predicting function}

\item{ind}{the rule to split the data to workers}

\item{...}{optional arguments to pred_fun}
}
\value{
A list of the estimated value of dependent variable and other output of the pred_fun
}
\description{
Use distributed methods to predict dependent variables based on estimated parameters
}
\examples{
# Not Run
}
